Create component to actions buttons (generate, save and auto-generate)
